applet & gui & transfer: improve error handling
This commit defines a way for the server to handle Glasgow/USB device errors. These errors occur when Glasgow is unplugged or reset, and are not recoverable without re-starting the server.  If reads and writes to the device return a GlasgowDeviceError or USBErrorOther, the server will write EOF to the client so that the client can stop waiting for a response that isn't coming.

Error handling has also been added to the image acquisition functions in the main GUI. If a TransferError occurs during live or photo acquisition, the acquisition process will stop and the GUI will reset its buttons to their initial states.